Split runner file dispatch tests
